requirements.txt
spotipy==2.16.1
Flask==1.1.2
requests==2.25.0
Pillow==8.0.1